Lift P9 true-anomaly geometry and favored-nu constants into core
