+++ crates/p9-2016-holman-payne/src/lib.rs
@@ -51,14 +51,17 @@ pub mod sky;
 
 use p9_core::types::P9Params;
 
-// Reuse the sibling's orbit and geometry verbatim — the two papers analyze the
-// same Brown & Batygin (2016) reference orbit and the same Sun–Saturn geometry.
-pub use p9_2016_cassini_ranging::brown_batygin_orbit;
-pub use p9_2016_cassini_ranging::geometry::{p9_position_at_true_anomaly, P9Geometry};
+// The two papers analyze the same Brown & Batygin (2016) reference orbit and
+// the same Sun–Saturn geometry; both now live in `p9_core`. The range-residual
+// physics is still reused from the sibling `p9_2016_cassini_ranging` crate.
+pub use p9_core::data::ephemeris_constraint::brown_batygin_orbit;
+pub use p9_core::types::{
+    position_at_true_anomaly as p9_position_at_true_anomaly, OrbitGeometry as P9Geometry,
+};
 
 /// Planet Nine on the Brown & Batygin (2016) orbit used throughout this paper.
-/// Alias of the sibling crate's [`brown_batygin_orbit`]; provided so callers can
-/// build the orbit without naming the sibling crate.
+/// Alias of the core [`brown_batygin_orbit`]; provided so callers can build the
+/// orbit without naming core.
 pub fn holman_payne_orbit() -> P9Params {
     brown_batygin_orbit()
 }
