+++ crates/p9-core/src/data/mod.rs
@@ -1,5 +1,6 @@
 //! Vetted observational data tables shared across the paper crates.
 
+pub mod ephemeris_constraint;
 pub mod etno;
 pub mod refresh;
 pub mod stable_kbos;
